vpnaas: install apparmor before restarting the service
This plugin is not used in neutron-vpnaas gate since Liberty where we
use a devstack plugin. Sadly, the code is still used in Kilo gate of the
project.

This change will need to be backported up to kilo.

The change reflects the corresponding patch in neutron-vpnaas devstack
plugin: Ic569c43e014ef11f8121db7dd5047a39579a4633

Change-Id: I6a2a127c3e17515172fbbd3bf1a1d1d1721e6cf6