Update patch set 1
Patch Set 1: Code-Review+1

This did hit our PowerVM CI and increased the log size a lot (because the term wasn't set by default).  It took a 2 meg log and made it about 8 megs.

We put a patch in to fix it on our side, but this might be nice for others.

Patch-set: 1
Label: Code-Review=+1