Update patch set 1
Patch Set 1:

Cool - I've been testing it out and am hitting some odd issues.  It is complaining about the Router's Gateway IP...

I'm wondering if that is a new value I need to provide.

I'm going to look through that first, then I'll take the bits out of stackrc and lib/neutron-legacy and re-propose.

Thanks for the feedback!

Patch-set: 1