Use Apache/mod_proxy as the TLS proxy instead of stud
Setting a proxy for a service is now a two-step process:

1. Add a configuration for the listener port using add_tls_proxy
2. Restart Apache using start_tls_proxy

Some services have multiple ports so this reduces the number
of Apache restarts required.

A template is used to generate configuration for each
service and is appended to tlsproxy.conf in the
appropriate Apache config subdir. This is still
readable and makes cleanup simpler.

No SSLCertificateChainFile/SSLCACertificateFile is set
because the devstack CA is added to the system bundle.

The SSL/TLS certificate generated by devstack is now
broken out into separate cert/key files for use by
Apache.

Move enable_mod_ssl from lib/tls to lib/apache where it belongs.

Add enable_mod_proxy function so Apache can act as a TLS proxy.

If the tlsproxy service is enabled then ensure that Apache,
mod_ssl and mod_proxy are all enabled. Don't rely on Keystone
to enable these.

Change-Id: Iafb2fa24b7738779c82171dfc57d43877c2e5ed6