Make wait_for_service more robust by checking HTTP response
wait_for_service just checked to see if the remote service
was started, not that it was returning data. This caused
problems when the service was behind a proxy because the
proxy would respond quickly but the service may not have
fully started.

Wait for either a 200, 300 or a 401 or timeout trying.

Change-Id: I059a12b1b920f703f28aca0e2f352714118dee97