Make wait_for_service more robust by checking HTTP response
wait_for_service just checked to see if the remote service
was started, not that it was returning data. This caused
problems when the service was behind a proxy because the
proxy would respond quickly but the service may not have
fully started.

Wait for a non-503 HTTP response code and non-7 exit code
(connection error) from curl

The \EOF form is used so the heredoc isn't evaulated before
the call is made. Otherwise it would be evaluated just once
so the loop is not valuable.

Variables needed to be exported in order to be available when
thsi form is used.

Change-Id: I059a12b1b920f703f28aca0e2f352714118dee97