Update patch set 3
Patch Set 3:

If I put on my operator hat, probe ports are useful constructs.  But if we aren't going to use them in the gate, then I question why we are allocating them...

Now, I'm not 100% sure whether I want to completely take them out of devstack (as is done here), or just change devstack-gate to not allocate them... thoughts?

Patch-set: 3