Update patch set 4
Patch Set 4:

So I don't think we've made a real decision on supporting MySQL -- I think it's just that this was done before the split and nobody has ever changed it.  It's mariadb on rpm platforms; I wonder if we should just unconditionally change it to mariadb for a) consistency and b) that seems to be the way forward into the future.

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0