Change flag name for test_volume_backup_export_import
Bug #1476416 was fixed in Liberty release. So the
test_volume_backup_export_import should be skipped
for Kilo gate. The patch changes the name
CONF.volume_feature_enabled.backup_import with
CONF.volume_feature_enabled.skip_backup_import. This
is done to avoid misunderstanding as we skip the
backup import test in kilo as the parameter is True.

Change-Id: I78aa9a1e01e829526ffcede0a216c9227c5cd09d
Related-Bug: #1476416