Have a way to disable Glance v1 in devstack
As part of the process of deprecating Glance's V1, the glance team would
like to start testing V2-only environments. Therefore, this change
provides a way to force other services to use V2.

This patch also adds a way to disable V1 entirely but this is currently
commented as not all services are able to talk to Glance's V2 yet. The
team will be switching other services as soon as the compatibility work
is completed.

Change-Id: I87e77d07964eac01e9a796817cbc88bd6e59c721