Update patch set 1
Patch Set 1: Code-Review-1

Let's find another way to express this then. We already have a couple hooks for things like security group

https://github.com/openstack-dev/devstack/blob/master/lib/neutron_plugins/ml2#L135

So let's make a new hook for if a plugin has l3 services.

This is why Q_L3_ENABLED was unhelpful - it had tons of meaning to different people.

Patch-set: 1