Update patch set 2
Patch Set 2:

> Is there a reasonable case where this should stay off?  I am not
 > familiar with the specifics of what changes in Swift here.  We like
 > to stick to app defaults whenever possible, and this can be set
 > from local.conf when necessary (except in the gate).

I added Tempest cases here (https://review.openstack.org/#/c/306237/) that relies on storage account management.   One of the review comments was to check this via configuration and that the gate configuration should enable all of the optional functionality.

During a conversation with Matt from QA on IRC, we proposed this change.   Right now, there is an issue with Swift functional tests if this is enabled with Keystone and I have a fix for the Swift functional tests here:  https://review.openstack.org/#/c/316303/.

Ultimately, the default tempest configuration needs to match the default devstack Swift configuration, so either we disable this in both or we enable it in both.

Patch-set: 2
Reviewer: Gerrit User 20553 <20553@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0