Allow setting libvirt images_type via env variable
Nova wants to change the postgres job to use 'raw' for the
image backend when using the libvirt (default) virt driver
for CI testing. By default, the images_type is based on
the 'use_cow_images' option which defaults to True, so the
default image backend is qcow2, and all other jobs, except
for ceph and parallels and lvm, are using that - which is
really most of the integrated gate since parallels is
third party CI and LVM is in the experimental queue.

So we want to get some additional coverage in our test matrix
and this change adds the ability to set CONF.libvirt.images_type
when using libvirt and not using parallels/ceph/lvm.

Related to nova blueprint libvirt-instance-storage

Change-Id: I5a31172b9a4fb86504adffd96bc466de22977df5