DNM: use libvirt.images_type=flat by default
We have test coverage already for the ploop, rbd, lvm and
qcow2 image types in the libvirt image backend.

We don't have a job that uses flat/raw (yet), so this is
just a one-off patch to test the refactor series in nova
with the flat backend.

Related to nova blueprint libvirt-instance-storage

Depends-On: I3281d32a87bac136ad32902a02bb9c9a39e4f9a4

Change-Id: I5a31172b9a4fb86504adffd96bc466de22977df5