DNM: use libvirt.images_type=flat by default
We have test coverage already for the ploop, rbd, lvm and
qcow2 image types in the libvirt image backend.

We don't have a job that uses flat/raw (yet), so this is
just a one-off patch to test the refactor series in nova
with the flat backend.

Related to nova blueprint libvirt-instance-storage

Depends-On: I70a559f3dc9b59097ff6923920f4377cca00d1b2

Change-Id: I5a31172b9a4fb86504adffd96bc466de22977df5