Switch fernet back as the default token provider
Commit I588a84c5179eab072d21bc1394aea2df00929650 reverted
181588b9ba50fefea12617a9a5e774daddfed20c because it was causing gate
failures and there was not a good progress on landing a fix for that.

The keystone team believes there was a bug in the token revocation
logic. Change Ib6a686494e897840b09d134ecf1ca50ce712f281 proposed a fix
for it.

This change will make it possible to test such fix proposal against
tempest jobs. A change will be submitted to tempest and there will be
consecutive job rechecks in order to try to catch the race condition.

This SHOULD NOT be merged until there is enough confidence the bug has
been fixed.

Related-Bug: #1577558
Related-Bug: #1578866

Change-Id: If2fbd4df845a020142e90af9777186298ec4750c