Update patch set 1
Patch Set 1:

so you're saying you want to do

 time_start 'foo'
 some_long_part1
 echo "That took $(time_elapsed 'foo')"
 some_long_part2
 time_stop 'foo'

I'm hesitant to just add functions that see little use ... why not just use two different names for part1 & 2?

The real advantage to these time_* fns was collating the summary table so you could quickly see where time was going at the end.  Probably a datestamp in between does just as much as this if you're manually walking the logs ...   

I admit this is a bit of a nit-pick, but this does become an api that remains for eternity so i like to be convinced.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0