Update patch set 8
Patch Set 8: Code-Review-1

Sorry to be annoying, but I'd just like to make sure we're fixing the right thing here

It seems like the original code path here does a stop & start on libvirt service for Xenial

Are you saying that this stop&start is not sufficient to get libvirt to recognise the new file on Xenial?  Because that sounds wrong and possibly points to a problem with the Xenial libvirt init files?

The restart as presented here seems fine as an optimisation, with the branch because of the known issue that "restart" doesn't work on Trusty.  But the changelog suggests it's fixing a bug...

Patch-set: 8
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1