Update patch set 1
Patch Set 1:

It will be used in wait_for_service for use with real proxies. A proxy will return a 503 when a service isn't available and wait_for_service doesn't currently handle this case at all. In order to handle this it needs to first wait for the requested port to become available and then wait for a non-503 result. Given this is two steps, in order to enforce a timeout for the function the time passed between events needs to be calculated.

Patch-set: 1