Update patch set 4
Patch Set 4:

IMO DevStack is a good start for OpenStack beginners. To test latest OpenStack sources, generally we run DevStack with master branch. However, for OpenStack users such as network engineers who want to test or integrate OpenStack with physical network devices, and users who want to see overall running components and instances, it is better to run DevStack with latest stable branch (now stable/mitaka). Some developers need to test OpenStack in new Linux distributions with a stable branch. Although users can run DevStack with FORCE=yes option, this option is not relevant for beginners.

Newbies tend to run DevStack with latest Linux distributions such as Ubuntu 16.04. I think it is better to include xenial although it is now in stable branch. AFAIK xenial works so well in stable/mitaka branch. This is why I have backported from the master branch.


If removing vivid in the stable branch is not good idea, then, is it okay just to include xenial in stable/mitaka branch?

Patch-set: 4