Update patch set 1
Patch Set 1:

I think given the flip to using transport_url we can both be satisfied with the outcome. I can handle the port change in the call to get_transport_url() and the rest is self-contained in the second block of code in my patch. Once https://review.openstack.org/#/c/320746 lands I'll update my patch with this new approach.

Patch-set: 1