Make wait_for_service more robust by checking HTTP response
wait_for_service used to just hit the url with curl
until a non-zero error was returned. This relied on the fact
that curl returned 7 on connection errors.

This didn't actually evaluate whether the service was up
and responding so it broke when checking for a service behind
a proxy. The proxy would return a 503 and this satisfied
the connection check but didn't assure that the service was up.

This adds an additional, looping test to ensure that the
server returns a non-503 answer.

Change-Id: I059a12b1b920f703f28aca0e2f352714118dee97