Update patch set 6
Patch Set 6: Code-Review-1

This is a question mostly because I haven't been following privsep - is it intended to replace rootwrap entirely in the future? Or only parts? Because with lib/neutron I've already switched rootwrap to daemon only and not supporting the old rootwrap command. With lib/neutron I really don't want to expand the matrix of things we need to configure and support.

Patch-set: 6
Label: Code-Review=-1