Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

You still have a style issue in the code.

IMHO this warning message created because we do not want this assignment to be there. If the CLI can operate without this, we should remove the assignment as well.

It is meant to be annoying, because this is what needs to be fixed.

So please fix the cli first, than remove the assignment.

Patch-set: 4
Label: Code-Review=-1