Update patch set 1
Patch Set 1:

Based on the commit message I would expect '[A\B]' -> 'A' not '[A\B]' -> ''.
You described A as the 'section name'.

Can you make at least this more clear in the message ?


This kind of usage does not seams generic enough to add it devstack,
the ironic change uses this function only once for processing a file which is very ironic helper script specific thing.

PS.: Also in the ini files perspective '\' '.' interpreting as `subsection` just higher level concept, the lower level function should handle everything in [] as one.

Patch-set: 1