Update patch set 5
Patch Set 5:

If the word (EOF in this case) is quoted then the expressions in the heredoc are not expanded, which is what I need because otherwise it is only expanded once. What you see in execution where the EOF is unquoted is something like:

+ timeout 10 bash -x
++ curl -g -g -f -k --noproxy '*' -s -o /dev/null -w '%{http_code}' https://localhost:9292/
+ [[ 503 = \5\0\3 ]]
+ sleep 1
+ [[ 503 = \5\0\3 ]]
+ sleep 1
...

The curl is executed exactly once, expanded when the heredoc is first evaluated.

If EOF is quoted and the variables aren't exported then for CURL_GET:

+ CURL_GET
test.sh: line 6: CURL_GET: command not found
++ -g -f -k --noproxy '*' -s -o /dev/null -w '%{http_code}'
bash: line 1: -g: command not found

or for url:

curl: no URL specified!
curl: try 'curl --help' or 'curl --manual' for more information

http://www.gnu.org/software/bash/manual/bash.html#Here-Documents

Patch-set: 5