Move tempest plugin install to the end
For the tempest plugin install inside the tox venv to hold we need to
ensure that it's the last thing run that touches the tox venv before
devstack ends. Otherwise there is a chance we'll recreate the venv in
a later step of installing and configuring tempest. This commit
moves the plugin installation into it's own function and calls that
function as part of post-extra phase of the tempest setup to make sure
it runs last.

Change-Id: Ie253171537e8c5a9887cc30aba1cad4b31e57663