Update patch set 4
Patch Set 4:

This has come up again in I8727938a938f2ae25aa098be2dc3003548c827ae

If we just want to add Xenial, then we can modify this to do that.  As I stated in previous reviews, I don't think anyone is going to maintain vivid so we might as well take it out.  or if people are totally against adding it to this branch, let's -2 this and state the policy somewhere.

Patch-set: 4