Update patch set 4
Patch Set 4:

Just a bit of context, it's only in the past two releases or so where QA resources were dedicated to the stable branches where it was actually possible to reliably run DevStack and have it actually complete, with zero errors, on a stable branch.

I remember for years, there'd be problems with newer versions of dependencies getting installed, or packages being deleted, etc. - eventually I would give up and just install master - because that was running at the gate and was the most maintained. Devstack's branches for a long time were just snapshots in time - and it was a flip of a coin if they worked or not. Obviously, with the changes made recently with requirements and constraints, the stable branches function more reliably, but honestly it's still the master branch where people should be doing work. If you absolutely must run a stable branch and insist on a distro, then read our warning and use FORCE - however you are ON YOUR OWN.

Patch-set: 4