Change the DEFAULT_IMAGE setting logic for ironic
Currently it is set depending on the driver, but it should not be
anymore, as both pxe_* and agent_* drivers support both whole disk
and partition images. Let's determine the DEFAULT_IMAGE based on the
IRONIC_TEMPEST_WHOLE_DISK_IMAGE parameter from localrc, and, when it
is not set, based on requested ephemeral partition size.

Depends-On: I4df45636f52a1ed08368e122160cc79b6c423af2
Change-Id: Ic4624d8aed9c89657c42f581d600551be627c761