Glance: setup cinder store for glance_store gate functional test
When GLANCE_STORES and GLANCE_DEFAULT_STORE environment variables
are set, this will setup the stores and default_store based on
them. Otherwise, it will enable stores based on enabled services,
as it has done so far.

This will be used to enable glance cinder store for the
glance_store's functional test job.

Change-Id: Iefe09e6915b4e073d6b95bca9c20fde33bf3af79
Needed-by: I20dec7ec28844bb03887917320a7d1da296e36d7