Update patch set 1
Patch Set 1:

> Is this really required? Especially when we print the warning when it occurs.

The warning says "this script has not been tested", but some DevStack users do not know what the warning means. Users want to know which situation can make the warning, and for developers the fact (not tested) is usually more helpful than describing the situation. So I wanted to explain that 'not been tested' could mean that 'the target OS is not supported' or 'the target OS is too new'.


I agree that it is not required, but adding such note will make more developers and potential users happy I think (potential OpenStack users might not be familiar with the situation).

Patch-set: 1
Label: Verified=0