Neutron: remove Q_ML2_PLUGIN_TYPE_DRIVERS
This option basically mimics the defaults, so rather than updating it
everytime the default in Neutron changes[1], and having to update
DevStack[2], let's just rely on the default.

If someone wants a different setting, they should place it in their
local.conf or have a DevStack plugin that configures what they need.

[1]: I8c29a1c1a7c79e02c26ac9e2ad2645d30dfbeefc
[2]: Id75651dfe57a07045a6932a0369668f33c7eef09
Change-Id: I7c10dfbc04cb11cb7792860bae13cd050d189f58