Update patch set 2
Patch Set 2:

(1 comment)

I think we can just fix it in here.

While there's no one answer, I think we should assume that non-interactive use is going to be captured to a plain text file and uploaded for presentation via HTTP from a gerrit message as in a 3rd party CI situation.  In this case, we don't want escape codes.

Patch-set: 2