Update patch set 3
Patch Set 3: -Code-Review

That all being said I thought about this more and changed it from a +2. I'm curious why we think this path is actually necessary. This is only for the case where the variable is not set to a comma separated list. This seems like just extra complexity added to devstack when we're essentially protecting against user misconfiguration. Why not change the docs to say it needs to be a comma separated list?

Patch-set: 3