Update patch set 3
Patch Set 3: Code-Review-1

Oh, I did some more digging and there is a bit more background to this change. It is here: https://review.openstack.org/#/c/304530/ which was missing a release note. Which was important for a change like that...

I think we need to update the commit message to better describe the situation. There is no history there, just saying it changed. (which I personally wasn't aware of) Also, this is coming out of order, we should have updated devstack before we removed anything from tempest.

That all being said I still am not convinced this is actually needed, when the simple fix is simply for users to use a comma separated list

Patch-set: 3
Label: Code-Review=-1