Update patch set 3
Patch Set 3:

Yes, please, this is needed to run existing test:
tempest/api/volume/admin/test_multi_backend
and will be needed for to-be-merged test:
https://review.openstack.org/#/c/326681/ Migrate volume while attached to an instance
(2nd test is vital for un-tested paths in cinder migrate as well as nova swap_volume)

I realized that a job will need to be configured to actually set >1 backends. There is currently a patch to get this started (experimental queue):
https://review.openstack.org/#/c/330678/
but we'd like this running for non-experimental, i.e. test_multi_backend (and probably test_migrate_volume_attached) would run today in gate-tempest-dsvm-full with >1 backend configured.

Patch-set: 3