Update patch set 3
Patch Set 3:

I understand the rationale of why. I think I outlined that in my comments. My -1 here is the commit message here does not explain why this changed and when it changed, and why devstack needs to be updated. All it does is outline the differences.

My other problem here, which is not blocking this, but needs an explanation in the commit msg at least, is why devstack is parsing the response at all. Why isn't this just a passthrough to tempest conf? Tempest wants a comma seperated string for the value of that config opt and the variable in devstack is a str. It seems to me that we don't need the parsing logic at all, and just do the iniset directly with the variable's value

I also would like to see something actually using this option, because clearly nothing has been testing it, otherwise when you blindly rmeoved the deprecated options from tempest nothing was using it in the gate to block that from landing. I'd like to avoid repeating the mistakes of the past

Patch-set: 3