Update patch set 2
Abandoned

The tempest tests that need this flag were removed from the patch set at this moment, there is a Glance bug open that is blocking those tests (bug 1588007). Once that bug is resolved I will include the tests in a different patch set and retake this patch set.

Patch-set: 2
Status: abandoned