Setsid to fix "old" screen_it--screen_stop chain
functions-common:screen_it utilizes _old_run_process in case of a
non-screen environment such as grenade testing.  However this call-chain
fails to store correct PID/PGID in order for a service to be stopped by
screen_stop later on.  The issue boils down to following (simplified) code:

    function _old_run_process() {
        local command=$2
        exec /bin/bash -c "$command"
    }

    _old_run_process service-name "cd workdir && \
                                   ./service --config service.conf" &
    echo $! > state/service.pid

The issue is two-fold. The pid of the sub-shell is being saved instead of
the service "$command" pid.  The screen_stop chain uses pkill -g (-F)
state/service.pid to terminate the whole service process group.

Since pid of the subshell is saved instead of the pgid of the service,
pkill -g fails with exit code 1 (target not found).  If a correct pgid was
saved instead (echo `ps -opgid= $!` > state/service.pid), due to pgid
implicit inheritance the pkill command would have killed the whole grenade
process group.

Suggested change utilizes setsid to generate new pgid (and sid) of the
service instead:

    function _old_run_process() {
        local command=$2
        exec setsid /bin/bash -c "$command"
    }

    _old_run_process service-name "cd workdir && \
                                   ./service --config service.conf" &
    echo $! > state/service.pid

Because exec is used here, setsid isn't forked as a separate process but
rather replaces subshell process instance while maintaining all
redirection.  Setsid uses its pid to create new pgid and sid of a
"$command" that ist's going to exec, again, without forking.  Since that
pid is stored in the parent shell ($!), it can be used as the pgid (and
sid) of the service process eventually forked by bash in the "$command".
This pgid being different from parent shell's, it is safe to pkill -g that
pgid to stop the service only (the forked service has ppid==1,
pgid==sid==$!, pid==<new pid>). That effectively turns the service into a
demon process.

Correctness of this change can be checked by executing pstree -pg instead
of "$command" to show the shell relationship as well as pid and pgid of
each process involved.

Change-Id: I265538ef9e9925ee64477ae127fd050e6133d578
Closes-Bug: 1595225