Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

This is sorely needed so positive in general.

If it needed to be done in sed ... well, here we are.  i have some half-written code for a python-based ini tool.  TBH i didn't go through the inner workings of it all -- i guess the hope is that basically nobody ever will.

I really think people are going to want to stick this into their jobs in various ways quickly, so I would *really* like to see a description of how to use this to achieve common tasks in the documentation and so we have an -- even very vague -- api reference for it.

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1