Enable plugins to determine if they want to provision L3 resources
Commit c35110e7c5c35dd1edc310dc3d0bb8693e58d336 added the ability
to determine if a plugin has extensions enabled and then provisions
the L3 resources. This is not good enough in the event that the plugin
requires additional hardware for this opertaion and that is not
avalaible.

This patch gives the option for the devstack deployment to specify
that they do not want the L3 resources created.

The default value is True.

Change-Id: I2073de07dcb7c4a47029872c0c2cf0be61850056