WIP Install a no-rules policy file for Nova
Nova has moved policy rules into code so that it is no longer necessary
to set policy rules in a config file. However oslo.policy requires a
policy file to exist, and it must be parseable JSON. This installs a
policy file with no rules defined that is still a valid JSON file.

Change-Id: Ib4abd1cf7a010f04e495b97776c36329c005111d