Parent:     8e64c478 (Merge "Tempest: add a Keystone reseller feature flag") Author:     Wenyan Zhang <zhang.wenyan3@zte.com.cn> AuthorDate: 2016-07-07 19:05:46 +0800 Commit:     Wenyan Zhang <zhang.wenyan3@zte.com.cn> CommitDate: 2016-07-07 19:05:46 +0800
Remove all *.pyc files in $DEST when executing clean.sh

Currentlly, the *.pyc files could not be removed in any scripts or
functions. But the redundant files would lead stack.sh not to find the
correct script for some versions after branch switched from master to
stable/mitaka in migration_helpers.sync_database_to_version.

So this commit adds the process of cleaning all the *.pyc files in
clean.sh.

It is needed to execute clean.sh before re-stack.sh to prevent the
exception.

Change-Id: If66a8db37d7ba51534b8303eca260dbf0b58073d
Closes-Bug: #1599124