Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

so i do get that it is confusing that the create-user script bails out with issues with the network ... but I don't see how this really fixes it.  lots of the other scripts source stackrc as the way to get the config variables in, and they will all fail in the same way?  the problem seems to be with the network configuration causing the die() to hit.  i'm not convinced this really helps other than move the problem to the next step

Patch-set: 2
Label: Code-Review=-1