Update patch set 2
Patch Set 2: Code-Review-1

So I'd just say that Id749c37ab7fefa96b35f11816b56b9def5ef4b08 is wrong, and uninstalling python-virtualenv is the wrong thing to do on all platforms.  we should just pip-install it (in fact, we should probably package_intsall python-virtualenv and *then* pip overwrite it).

At any point later we could have dependencies that end up re-installing this package and making another mess of it.

BTW, the bug is a better commit msg, it gives context as to what's going on

Patch-set: 2
Label: Code-Review=-1