Update patch set 1
Patch Set 1: Code-Review-1

we sure this is idempotent?  because now for the usual all-in-one case we do this twice?  that seems to suck  

so nova depends on os-brick ... is that common enough to both that we can just do it once in there?  where else could we do it once (maybe just stack.sh?)

Patch-set: 1
Label: Code-Review=-1