Always set external network type as flat
Since we are currently using the 'external_network_bridge' option
for the L3 agent, the type of the external network doesn't matter
too much currently. However, it impacts the calculation of the MTU
value that the router interface will use once we stop using
the deprecated 'network_device_mtu' option, so we need to make sure
the external network type is 'flat'.

Setting the provider type was conditional on the
Q_USE_PROVIDERNET_FOR_PUBLIC flag, so we could conceivably adjust
all of our gate jobs to use that. But there doesn't seem
to be a reason not to just always set the external network type
to 'flat' instead. An external network of type 'vxlan' would be
unreachable by the host node anyway.

Change-Id: Id20e67aba5dfd2044b82c700f41c6e648b529430
Related-Bug: #1603493