Update patch set 2
Patch Set 2:

YAMAMOTO, ok, I was a bit wrong, but I think the general principle stands: we should not load the whole directory, we should pick specific files per service. The fact that dhcp_agent.ini won't be loaded is a mere implementation detail. Other repos may ship a .conf file designated for a service that is not neutron-server, and then you will attempt to load it.

Patch-set: 2