Update patch set 1
Patch Set 1:

Apache is already listening on 443 in ports.conf by default.

Keystone is moving away from ports to use URLs instead so most of the requests within devstack look like curl --options https://localhost/identity_v2_admin

As for SSLEngine, sure I could hardcode it but I felt that it would be clearer without. Right now with USE_SSL=False this shows up in keystone.conf:

#<VirtualHost *:443>
#    
#    
#    
#</VirtualHost>

The first line in the VH would otherwise be:

#    SSLEngine on

I'm not opposed to changing this if you still want, but this was my reasoning.

Patch-set: 1