Fixes for linux bridge and Q_USE_PROVIDER_NET
===Make move address function check for OVS===
This function is specific to OVS however its called from the L3
code regardless of plugin type. This patch adjusts the logic
to specifically check if OVS is used before calling it.

===Set bridge_mappings for linux bridge===
The external network physnet needs a bridge_mapping to the public
bridge when the L2 agent is responsible for wiring.

===Add PUBLIC_PHYSICAL_NETWORK to flat_networks===
This network must be present in the ML2 flat_networks config if
flat_networks is specified.

Closes-Bug: #1605423
Change-Id: I95d63f8dfd21499c599d425678bf5327b599efcc