Update patch set 2
Patch Set 2: Code-Review-1

I'm not going to deny something is wrong, but even after a fair bit of research here it's not clear to me what it is.

The change that is referenced [1] and this variable are talking about easy-install.pth -- we use pip everywhere (see [2]).  AFAIK pip does not do the path munging stuff or even have that file?

When we setup a package from git, that *has* to be an equal or later version than what was installed via pypi?  pip not recognising this seems like a serious issue.

I'm trying to get some sort of minimal reproduction of this issue inside a virtualenv, but haven't had luck with that so far.  This would certainly help us understand this to root-cause the problem.

I am fine with workarounds ... once we understand what's actually going on.  So far pip doesn't seem to have any issues open on it [3]; that does suggest maybe it is something specific to devstack?

[1] https://github.com/pypa/setuptools/commit/e211f4584ac59f404a3e1a17c6b54dff18468cf6
[2] https://git.openstack.org/cgit/openstack-dev/devstack/tree/inc/python
[3] https://github.com/pypa/pip/issues

Patch-set: 2
Label: Code-Review=-1